refactor(write_engine): split cql_to_mutation into builders/codec/delta_helpers (#1123)
Behavior-preserving module split of the ~3,594-line write-path file
cql_to_mutation.rs into a directory module with one concern per file:

- cql_to_mutation/mod.rs       — string-dispatch entry points
  (convert_cql_to_mutation / convert_cql_to_mutations) + pub(crate)
  re-exports of the unchanged public surface (literal_to_value and the
  INSERT/UPDATE/DELETE/BATCH builders)
- cql_to_mutation/builders.rs  — insert/update/delete/batch → Mutation
- cql_to_mutation/codec.rs     — literal/expression/JSON → Value coercion,
  scalar parsers (uuid/blob/inet/varint)
- cql_to_mutation/delta_helpers.rs — WHERE bindings, predicate extraction,
  range tombstones, table validation, USING TIMESTAMP/TTL

No semantic changes: every production function body is identical; the only
diffs are rustfmt signature re-wrapping and visibility (pub(super) for
cross-submodule helpers, pub(crate) preserved for the re-exported surface).
Tests follow their code into per-concern #[cfg(test)] mod tests; shared
schema/WHERE fixtures are relocated next to the code that uses them. Test
count preserved (98 cql_to_mutation lib tests, all passing).

Parent write_engine/mod.rs is untouched (its `mod cql_to_mutation;` line
auto-resolves to the new mod.rs), avoiding any conflict with #1120.
